updates per pr & some cleanup
