+++ bin/cli.js
@@ -3,6 +3,6 @@
 
 require('codemod-cli').runTransform(
   __dirname,
-  process.argv[2] /* transform name */,
-  process.argv.slice(3) /* paths or globs */
+  'convert-htmlbars-import' /* transform name */,
+  process.argv.slice(2) /* paths or globs */
 );
